move card actions logic to own file
