+++ web/src/app/details/DetailsPage.tsx
@@ -1,11 +1,9 @@
-import React, { MouseEventHandler, ReactNode } from 'react'
+import React, { ReactNode } from 'react'
 import statusStyles from '../util/statusStyles'
 import { makeStyles } from '@material-ui/core/styles'
 import Card from '@material-ui/core/Card'
 import CardHeader from '@material-ui/core/CardHeader'
 import CardContent from '@material-ui/core/CardContent'
-import CardActions from '@material-ui/core/CardActions'
-import Button from '@material-ui/core/Button'
 import Grid from '@material-ui/core/Grid'
 import Typography from '@material-ui/core/Typography'
 import { ChevronRight } from '@material-ui/icons'
@@ -14,10 +12,12 @@ import ListItem from '@material-ui/core/ListItem'
 import ListItemText from '@material-ui/core/ListItemText'
 import ListItemSecondaryAction from '@material-ui/core/ListItemSecondaryAction'
 import IconButton from '@material-ui/core/IconButton'
+
 import Notices, { Notice } from './Notices'
 import Markdown from '../util/Markdown'
 import AppLink from '../util/AppLink'
 import useWidth from '../util/useWidth'
+import CardActions, { Action } from './CardActions'
 
 interface DetailsPageProps {
   title: string
@@ -47,14 +47,6 @@ type Link = {
   status?: LinkStatus
 }
 
-type Action = {
-  label: string
-  handleOnClick: MouseEventHandler<HTMLButtonElement>
-
-  icon?: JSX.Element // if true, adds a start icon to a button with text
-  secondary?: boolean // if true, renders right-aligned as an icon button
-}
-
 function isDesktopMode(width: string): boolean {
   return width === 'md' || width === 'lg' || width === 'xl'
 }
@@ -65,29 +57,10 @@ const useStyles = makeStyles({
     height: '100%', // align with quick links if shorter in height
     position: 'relative', // allows card actions to remain at bottom, if height is stretched
   },
-  cardActions: {
-    // height: '100%',
-    alignItems: 'flex-end', // aligns icon buttons to bottom of container
-
-    // moves card actions to bottom if height is stretched
-    position: 'absolute',
-    bottom: '0',
-    width: '-webkit-fill-available',
   titleFooterContent: {
     paddingTop: 0,
   },
 })
-
-function Action(p: Action): JSX.Element {
-  if (p.secondary && p.icon) {
-    return <IconButton onClick={p.handleOnClick}>{p.icon}</IconButton>
-  }
-  return (
-    <Button onClick={p.handleOnClick} startIcon={p.icon}>
-      {p.label}
-    </Button>
-  )
-}
 
 export default function DetailsPage(p: DetailsPageProps): JSX.Element {
   const classes = useStyles()
@@ -98,30 +71,6 @@ export default function DetailsPage(p: DetailsPageProps): JSX.Element {
     if (status === 'warn') return classes.statusWarning
     if (status === 'err') return classes.statusError
     return classes.noStatus
-  }
-
-  const action = (action: Action | JSX.Element, key: string): JSX.Element => {
-    if ('label' in action && 'handleOnClick' in action) {
-      return <Action key={key} {...action} />
-    }
-    return action
-  }
-
-  const renderActions = (): JSX.Element => {
-    let actions: Array<JSX.Element> = []
-    if (p.primaryActions) {
-      actions = p.primaryActions.map((a, i) => action(a, 'primary' + i))
-    }
-    if (p.secondaryActions) {
-      actions = [
-        ...actions,
-        <div key='actions-margin' style={{ margin: '0 auto' }} />,
-        ...p.secondaryActions.map((a, i) =>
-          action({ ...a, secondary: true }, 'secondary' + i),
-        ),
-      ]
-    }
-    return <CardActions className={classes.cardActions}>{actions}</CardActions>
   }
 
   return (
@@ -165,7 +114,10 @@ export default function DetailsPage(p: DetailsPageProps): JSX.Element {
               </CardContent>
             )}
 
-            {(p.primaryActions || p.secondaryActions) && renderActions()}
+            <CardActions
+              primaryActions={p.primaryActions}
+              secondaryActions={p.secondaryActions}
+            />
           </Card>
         </Grid>
